+++ src/sprite/sprite-billboard-type.ts
@@ -5,9 +5,9 @@ export enum SpriteBillboardType {
   /**
    * Sprite will be rotated towards the viewer on both the x-plane and y-plane.
    */
-  Spherical = "spherical",
+  spherical = "spherical",
   /**
    * Sprite will be rotated towards the viewer on the y-plane.
    */
-  Cylindrical = "cylindrical",
+  cylindrical = "cylindrical",
 }